修改ReportCourse
修改ReportCourse

+++ e-Look/src/main/java/com/e_Look/ReportCourse/model/ReportCourseService.java
@@ -27,7 +27,7 @@ public class ReportCourseService {
 			dao.insert(reportCourseVO);
 	}
 
-	public String getJSON(Integer status) {
+	public String getJSON(byte status) {
 		return dao.getJSON(status);
 	}
