+++ e-Look/src/main/java/com/e_Look/ReportCourse/model/ReportCourseService_JNDI.java
@@ -22,7 +22,7 @@ public class ReportCourseService_JNDI {
 			dao.insert(reportCourseVO);
 	}
 
-	public String getJSON(Integer status) {
+	public String getJSON(byte status) {
 		return dao.getJSON(status);
 	}
